[DMS-1100] Bump schema packages to 1.0.327 for bigint education org IDs (#890)
* [DMS-1100] Bump schema packages to 1.0.327 for bigint education org IDs

Upstream MetaEd fix adds format annotations to jsonSchemaForInsert integer
properties. Education organization IDs (schoolId, educationOrganizationId,
localEducationAgencyId, etc.) now emit as bigint instead of integer/int.

- Bump EdFi.DataStandard52/Homograph/Sample/TPDM.ApiSchema to 1.0.327
- Update authoritative fixture inputs from new packages
- Regenerate all golden test expected outputs


* Update provisioned schema integration test manifests for bigint columns


* Fix 9 remaining MSSQL columns missed by DDL cross-reference

Reference-flattened columns (e.g. CourseOffering_SchoolReferenceSchoolId)
had names not present in the DDL column name set.


* Regenerate provisioned schema integration test manifests

PostgreSQL manifest regenerated from actual database introspection.
MSSQL manifest updated with new StudentAcademicRecordExtension table,
columns, and constraints from the 1.0.327 sample schema package.


* Regenerate MSSQL provisioned schema manifest from real database

Previous attempts used scripted int->bigint replacements which missed
structural changes (new tables, view definitions). This manifest was
regenerated by running the integration test against a local MSSQL
instance with UPDATE_GOLDENS=1.


* Fix trailing newlines in MSSQL manifest view/trigger definitions

Local SQL Server appends an extra trailing newline to view and trigger
definitions compared to CI's containerized MSSQL 2022. Strip trailing
newlines from definition strings to match CI output.


* Fix MSSQL manifest view/trigger definition endings to match CI

Local MSSQL appends extra trailing newline: ;

Previous fix over-stripped to: ;
CI expects: ;


* Regenerate MSSQL manifest from local DB, fix trailing newline diffs

Generated from local MSSQL then adjusted for CI MSSQL 2022 differences:
- View definitions: local produces ;

, CI produces ;

- Trigger definitions: local produces END;
, CI produces END;


* Fix PostgreSQL smoke test scope for updated sample schema

The 1.0.327 sample schema changed the extension scope path for
schoolDistricts from $._ext.sample.addresses[*] to $.addresses[*].


* Fix smoke test credential secret to meet 32-char minimum

The default SysAdminSecret in Get-SmokeTestCredentials was only 12
characters, failing the Config Service client secret validation which
requires 32-128 characters. Use the same default as Add-CmsClient.


* Update remaining 1.0.326 refs to 1.0.327, fix bigint types in resolver fixture, add boundary tests

* Fix bigint types for EdOrg IDs in PowerShell scripts and ParentKeyPart columns in relational model

* Regenerate sample mappingset manifest after rebase onto main

---------

+++ src/dms/backend/EdFi.DataManagementService.Backend.RelationalModel.Tests.Unit/RelationalModelSharedHelperTests.cs
@@ -117,7 +117,7 @@ public class Given_Relational_Model_Shared_Helpers
                 ("DocumentId", new RelationalScalarType(ScalarKind.Int64)),
                 ("CollectionItemId", new RelationalScalarType(ScalarKind.Int64)),
                 ("ParentCollectionItemId", new RelationalScalarType(ScalarKind.Int64)),
-                ("SchoolId", new RelationalScalarType(ScalarKind.Int32)),
+                ("SchoolId", new RelationalScalarType(ScalarKind.Int64)),
                 ("Ordinal", new RelationalScalarType(ScalarKind.Int32))
             );
     }
